+++ src/core/QRDot.ts
@@ -1,3 +1,4 @@
+import { QRCanvas } from './QRCanvas'
 import {
   BasicFigureDrawArgsCanvas,
   DrawDotArgsCanvas,
@@ -23,6 +24,7 @@ const dotTypes = {
   fluid: 'fluid',
   fluidLine: 'fluid-line',
   stripe: 'stripe',
+  stripeRow: 'stripe-row',
   stripeColumn: 'stripe-column'
 } as DotTypes
 
@@ -41,7 +43,14 @@ export default class QRDot {
     this.dotSize = options.dotSize
   }
 
-  draw(x: number, y: number, getNeighbor: GetNeighbor): void {
+  draw(
+    x: number,
+    y: number,
+    getNeighbor: GetNeighbor,
+    qrCanvas: QRCanvas,
+    i: number,
+    j: number
+  ): void {
     const context = this._context
     const type = this._type
     let drawFunction
@@ -73,6 +82,9 @@ export default class QRDot {
       case dotTypes.stripe:
         drawFunction = this._drawStripe
         break
+      case dotTypes.stripeRow:
+        drawFunction = this._drawStripeRow
+        break
       case dotTypes.stripeColumn:
         drawFunction = this._drawStripeColumn
         break
@@ -82,7 +94,16 @@ export default class QRDot {
         break
     }
 
-    drawFunction.call(this, { x, y, size: this.dotSize, context, getNeighbor })
+    drawFunction.call(this, {
+      x,
+      y,
+      size: this.dotSize,
+      context,
+      getNeighbor,
+      qrCanvas,
+      i,
+      j
+    })
   }
 
   _drawSquare({ x, y, size, context }: DrawArgsCanvas) {
@@ -124,7 +145,7 @@ export default class QRDot {
     const cy = y + size / 2
     const originX = -size / 2
     context.translate(cx, cy)
-    canvasRoundRect(context)(originX, originX, size,size, size / 4)
+    canvasRoundRect(context)(originX, originX, size, size, size / 4)
     context.fill()
     context.translate(-cx, -cy)
   }
@@ -168,10 +189,7 @@ export default class QRDot {
   _drawFluidLine(args: DrawArgsCanvas) {
     this._drawFluid(args, true)
   }
-  _drawFluid(
-    { x, y, size, context, getNeighbor }: DrawArgsCanvas,
-    line = false
-  ) {
+  _drawFluid({ x, y, size, context, getNeighbor }: DrawArgsCanvas, line = false) {
     let roundedCorners = [false, false, false, false] // top-left, top-right, bottom-right, bottom-left
     if (!getNeighbor(0, -1) && !getNeighbor(-1, 0)) roundedCorners[0] = true
     if (!getNeighbor(1, 0) && !getNeighbor(0, -1)) roundedCorners[1] = true
@@ -184,76 +202,121 @@ export default class QRDot {
     context.arc(0, 0, size / 2, 0, 2 * Math.PI, false)
     context.closePath()
     context.fill()
-    if (!roundedCorners[0])
-      context.fillRect(-size / 2, -size / 2, size / 2, size / 2)
+    if (!roundedCorners[0]) context.fillRect(-size / 2, -size / 2, size / 2, size / 2)
     if (!roundedCorners[1]) context.fillRect(0, -size / 2, size / 2, size / 2)
     if (!roundedCorners[2]) context.fillRect(0, 0, size / 2, size / 2)
     if (!roundedCorners[3]) context.fillRect(-size / 2, 0, size / 2, size / 2)
 
-    if (line) {
-      const originLinWidth = context.lineWidth
+    if (line && !getNeighbor(0, 1)) {
       if (getNeighbor(-1, 1)) {
         context.beginPath()
-        context.lineWidth = size / 4
-        context.moveTo(0, 0)
-        context.lineTo(-size, size)
-        context.stroke()
+        context.arc(-size, 0, size / 2, 0, 0.5 * Math.PI, false)
+        context.arc(0, size, size / 2, Math.PI, 1.5 * Math.PI, false)
         context.closePath()
+        context.stroke()
+        context.fill()
       }
       if (getNeighbor(1, 1)) {
         context.beginPath()
-        context.lineWidth = size / 4
-        context.moveTo(0, 0)
-        context.lineTo(size, size)
-        context.stroke()
+        context.arc(size, 0, size / 2, 0.5 * Math.PI, Math.PI, false)
+        context.arc(0, size, size / 2, 1.5 * Math.PI, 0, false)
         context.closePath()
+        context.stroke()
+        context.fill()
       }
-      context.lineWidth = originLinWidth
     }
     context.translate(-cx, -cy)
   }
 
   _drawStripeColumn(args: DrawArgsCanvas) {
-    this._drawStripe(args, 'column')
+    return this._drawStripe(args, 'column')
+  }
+
+  _drawStripeRow(args: DrawArgsCanvas) {
+    return this._drawStripe(args, 'row')
   }
 
   _drawStripe(
-    { x, y, size, context, getNeighbor }: DrawArgsCanvas,
-    type: 'row' | 'column' = 'row'
+    { x, y, size, context, qrCanvas, i, j }: DrawArgsCanvas,
+    type: 'row' | 'column' | 'default' = 'default'
   ) {
-    const cx = x + size / 2
-    const cy = y + size / 2
-    context.translate(cx, cy)
-    context.beginPath()
-    context.arc(0, 0, size / 4, 0, 2 * Math.PI, false)
-    context.fill()
-    context.closePath()
-    if (type === 'row') {
-      if (getNeighbor(-1, 0)) {
-        context.fillRect(-size / 2, -size / 4, size / 2, size / 2)
-      }
-      if (getNeighbor(1, 0)) {
-        context.fillRect(0, -size / 4, size / 2, size / 2)
-      }
-    } else if (type === 'column') {
-      if (getNeighbor(0, -1)) {
-        context.fillRect(-size / 4, -size / 2, size / 2, size / 2)
-      }
-      if (getNeighbor(0, 1)) {
-        context.fillRect(-size / 4, 0, size / 2, size / 2)
+
+    function setRangeDisabled(width, height) {
+      for (let i1 = i; i1 < i + width; i1++) {
+        for (let j1 = j; j1 < j + height; j1++) {
+          qrCanvas.setDisabled(i1, j1)
+        }
       }
     }
-    context.translate(-cx, -cy)
+
+    function getRangeTrue(width, height) {
+      let isTrue = true
+      for (let i1 = i; i1 < i + width; i1++) {
+        for (let j1 = j; j1 < j + height; j1++) {
+          if (!qrCanvas.isDark(i1, j1)) {
+            isTrue = false
+            break
+          }
+        }
+        if (!isTrue) {
+          break
+        }
+      }
+      return isTrue
+    }
+
+    function drawItem(width, height) {
+      const cx = x + size / 2
+      const cy = y + size / 2
+      context.translate(cx, cy)
+      context.beginPath()
+      if (width === 1 && height === 1) {
+        // 画圆点
+        context.arc(0, 0, size / 4, 0, 2 * Math.PI, false)
+      } else if (width > 1) {
+        // 画横
+        context.arc(0, 0, size / 4, 0.5 * Math.PI, 1.5 * Math.PI, false)
+        context.arc(size * (width - 1), 0, size / 4, 1.5 * Math.PI, 0.5 * Math.PI, false)
+      } else if (height > 1) {
+        // 画竖
+        context.arc(0, 0, size / 4, Math.PI, 2 * Math.PI, false)
+        context.arc(0, size * (height - 1), size / 4, 0, Math.PI, false)
+      }
+      context.fill()
+      context.closePath()
+      context.translate(-cx, -cy)
+      setRangeDisabled(width, height)
+    }
+
+    const array =
+      type === 'row'
+        ? [
+            [3, 1],
+            [2, 1],
+            [1, 1]
+          ]
+        : type === 'column'
+        ? [
+            [1, 3],
+            [1, 2],
+            [1, 1]
+          ]
+        : [
+            [3, 1],
+            [1, 3],
+            [2, 1],
+            [1, 2],
+            [1, 1]
+          ]
+
+    array.forEach((comb) => {
+      if (getRangeTrue(comb[0], comb[1])) {
+        drawItem(comb[0], comb[1])
+      }
+    })
   }
 
-  _rotateFigure({
-    x,
-    y,
-    size,
-    context,
-    rotation = 0,
-    draw
-  }: RotateFigureArgsCanvas) {
+  _rotateFigure({ x, y, size, context, rotation = 0, draw }: RotateFigureArgsCanvas) {
     const cx = x + size / 2
     const cy = y + size / 2
     context.translate(cx, cy)